Consolidate upload/manager.ts into one sync file
