+++ glean/src/core/upload/manager.ts
@@ -2,31 +2,43 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import type { Configuration } from "../../config.js";
-import type { PingInternalRepresentation } from "../../pings/database/shared.js";
-import type PingsDatabase from "../../pings/database/async.js";
-import type { UploadResult } from "../uploader.js";
-import type { UploadTask } from "../task.js";
-import type { IPingUploadManager, QueuedPing } from "./shared.js";
+import type { Configuration } from "../config.js";
+import type { PingInternalRepresentation } from "../pings/database.js";
+import type PingsDatabase from "../pings/database.js";
+import type { UploadResult } from "./uploader.js";
+import type { UploadTask } from "./task.js";
 
-import { Context } from "../../context.js";
-import log, { LoggingLevel } from "../../log.js";
-import { isDeletionRequest } from "../../pings/database/shared.js";
-import RateLimiter, { RateLimiterState } from "../rate_limiter.js";
-import { UploadResultStatus } from "../uploader.js";
-import PingUploadWorker from "../worker/async.js";
-import Policy from "../policy.js";
-import uploadTaskFactory, { UploadTaskTypes } from "../task.js";
-import { UPLOAD_MANAGER_LOG_TAG } from "./shared.js";
+import { Context } from "../context.js";
+import log, { LoggingLevel } from "../log.js";
+import { isDeletionRequest } from "../pings/database.js";
+import RateLimiter, { RateLimiterState } from "./rate_limiter.js";
+import { UploadResultStatus } from "./uploader.js";
+import PingUploadWorker from "./worker.js";
+import Policy from "./policy.js";
+import uploadTaskFactory, { UploadTaskTypes } from "./task.js";
 
-// See `IPingUploadManager` for method documentation.
-class PingUploadManager implements IPingUploadManager {
+const UPLOAD_MANAGER_LOG_TAG = "core.Upload.PingUploadManager";
+
+export interface QueuedPing extends PingInternalRepresentation {
+  // The UUID identifier for this ping.
+  readonly identifier: string;
+}
+
+/**
+ * A ping upload manager. Manages a queue of pending pings to upload.
+ *
+ * Observes the pings database
+ * and whenever that is updated the newly recorded ping is enqueued.
+ */
+class PingUploadManager {
   // A FIFO queue storing a `QueuedPing` for each pending ping.
   private queue: QueuedPing[];
+
   // A set of the identifiers of pings being processed
   // i.e. pings that were removed from the queue by calling `getUploadTask`,
   // but have not yet been deleted from the database / re-enqueued by calling `processPingUploadResponse`.
   private processing: Set<string>;
+
   // A worker that will take care of actually uploading pings.
   private worker: PingUploadWorker;
 
@@ -56,6 +68,11 @@ class PingUploadManager implements IPingUploadManager {
   }
 
   /// PUBLIC ///
+  /**
+   * Get the next `UploadTask`.
+   *
+   * @returns The next upload task.
+   */
   getUploadTask(): UploadTask {
     const nextTask = this.getUploadTaskInternal();
 
@@ -70,9 +87,48 @@ class PingUploadManager implements IPingUploadManager {
     return nextTask;
   }
 
-  async processPingUploadResponse(ping: QueuedPing, response: UploadResult): Promise<void> {
+  /**
+   * Processes the response from an attempt to upload a ping.
+   *
+   * Based on the HTTP status of said response,
+   * the possible outcomes are:
+   *
+   * 200 - 299 Success**
+   *   Any status on the 2XX range is considered a successful upload,
+   *   which means the corresponding ping file can be deleted.
+   *
+   *   _Known 2XX status:_
+   *   200 - OK. Request accepted into the pipeline.
+   *
+   * 400 - 499 Unrecoverable error**
+   *   Any status on the 4XX range means something our client did is not correct.
+   *   It is unlikely that the client is going to recover from this by retrying,
+   *   so in this case the corresponding ping file can also be deleted.
+   *
+   *   _Known 4XX status:_
+   *   404 - not found - POST/PUT to an unknown namespace
+   *   405 - wrong request type (anything other than POST/PUT)
+   *   411 - missing content-length header
+   *   413 - request body too large Note that if we have badly-behaved clients that
+   *           retry on 4XX, we should send back 202 on body/path too long).
+   *   414 - request path too long (See above)
+   *
+   * Any other error**
+   *   For any other error, a warning is logged and the ping is re-enqueued.
+   *
+   *   _Known other errors:_
+   *   500 - internal error
+   *
+   * @param ping The ping that was just uploaded.
+   * @param response The response of a ping upload attempt.
+   * @returns Whether or not to retry the upload attempt.
+   */
+  processPingUploadResponse(ping: QueuedPing, response: UploadResult): void {
     const { identifier } = ping;
-    this.processing.delete(identifier);
+
+    if (this.processing.has(identifier)) {
+      this.processing.delete(identifier);
+    }
 
     const { status, result } = response;
     if (status && status >= 200 && status < 300) {
@@ -81,7 +137,8 @@ class PingUploadManager implements IPingUploadManager {
         `Ping ${identifier} successfully sent ${status}.`,
         LoggingLevel.Info
       );
-      await this.pingsDatabase.deletePing(identifier);
+
+      this.pingsDatabase.deletePing(identifier);
       return;
     }
 
@@ -91,12 +148,11 @@ class PingUploadManager implements IPingUploadManager {
     ) {
       log(
         UPLOAD_MANAGER_LOG_TAG,
-        `Unrecoverable upload failure while attempting to send ping ${identifier}. Error was: ${
-          status ?? "no status"
+        `Unrecoverable upload failure while attempting to send ping ${identifier}. Error was: ${status ?? "no status"
         }.`,
         LoggingLevel.Warn
       );
-      await this.pingsDatabase.deletePing(identifier);
+      this.pingsDatabase.deletePing(identifier);
       return;
     }
 
@@ -112,17 +168,56 @@ class PingUploadManager implements IPingUploadManager {
     this.enqueuePing(ping);
   }
 
-  async clearPendingPingsQueue(): Promise<void> {
+  /**
+   * Clears the pending pings queue.
+   *
+   * # Important
+   *
+   * This will _drop_ pending pings still enqueued.
+   * Only the `deletion-request` ping will still be processed.
+   *
+   * @returns A promise which resolves once the clearing is complete
+   *          and all upload attempts have been exhausted.
+   */
+  clearPendingPingsQueue(): void {
     this.queue = this.queue.filter((ping) => isDeletionRequest(ping));
-    await this.blockOnOngoingUploads();
   }
 
-  async blockOnOngoingUploads(): Promise<void> {
-    return this.worker.blockOnCurrentJob();
+  /**
+   * Block uploading during testing.
+   */
+  blockUploads(): void {
+    this.worker.blockUploading();
   }
 
+  /**
+   * Resume uploading during testing.
+   */
+  resumeUploads(): void {
+    this.worker.resumeUploading();
+  }
+
+  /**
+   * Enqueues a new ping and trigger uploading of enqueued pings.
+   *
+   * This function is called by the PingDatabase every time a new ping is added to the database
+   * or when the database is being scanned.
+   *
+   * @param identifier The id of the ping that was just recorded.
+   * @param ping An object containing the newly recorded ping path, payload and optionally headers.
+   */
   update(identifier: string, ping: PingInternalRepresentation): void {
     this.enqueuePing({ identifier, ...ping });
+
+    // Since everything is happening synchronously, we cannot await the upload
+    // response before deleting. We delete before the ping is uploaded.
+    //
+    // `isBlocking` will only ever be true during testing after calling `this.worker.blockUploading()`.
+    if (!this.worker.isBlocking) {
+      this.processing.delete(identifier);
+      this.pingsDatabase.deletePing(identifier);
+    }
+
     this.worker.work(
       () => this.getUploadTask(),
       (ping: QueuedPing, result: UploadResult) => this.processPingUploadResponse(ping, result)
